fix: refactor and move the files
